Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 int main(){
-    double a, b;
+    int a, b;
     char c;
-    scanf("%lf %lf %c",&a,&b,&c);
+    scanf("%d %d %c",&a,&b,&c);
     if (c == '+')
        printf("%d",a+b);
     else if (c == '-')
